move SfdxWorkspaceLwcTestRunnerInstallationChecker into run actions; dipose result file system watcher after timeout
